+++ CHANGES.md
@@ -1,3 +1,19 @@
+#v 1.3.0	5-07-2013
+##CHANGES
++ now also working on mac host with default JRE
++ working with java 6
+
+##Changed Files
+	modified:   src/main/BrainfuckingIDE.java
+	modified:   src/ui/FrameListener.java
+	modified:   src/ui/MainFrame.java
+	modified:   src/ui/Terminal.java
+	modified:   src/ui/UI.java
+	modified:   src/ui/extras/TextGenerator.java
+	modified:	CHANGES.md
+	modified:	README.md
+	
+
 #v 1.2.0	2-28-2013
 ##NEW:
 + debugging